Fix multiple code issues found in Mu2eG4 review
- PhysicalVolumeHelper.cc: Fix typo "intialized" -> "initialized"
- checkForOverlaps.cc: Fix type mismatch (G4double for int config values)
- finishNesting.cc: Modernize null pointer style (0x0/0 -> nullptr)
- Mu2eG4MT_module.cc: Fix unreachable else-if branch (dead code)
- Mu2eG4MT_module.cc: Fix memory leak (.release() -> .reset())
- SensitiveDetectorHelper.cc: Remove unused inner loop in filterTrackerStepPoints
- ExtMonFNALPixelSD.hh: Initialize pointers to nullptr
- TrackerPlaneSupportSD.cc: Comment out unreachable debug check
- PhysicsProcessInfo.cc: Fix potential int overflow (int -> size_t)
- Mu2eG4UserHelpers.cc: Add null check after dynamic_cast

Agent-Logs-Url: https://github.com/Mu2e/Offline/sessions/b50a9949-a25c-4a8b-b89d-b4e1d9af8a7f

+++ Mu2eG4/src/PhysicsProcessInfo.cc
@@ -189,7 +189,7 @@ namespace mu2e{
        << _allProcesses.size() << " "
        << endl;
 
-    int tcsum(0);
+    size_t tcsum(0);
     for ( map_type::const_iterator i=_allProcesses.begin();
           i != _allProcesses.end(); ++i ){
       ProcInfo const& oi        = i->second;
@@ -218,7 +218,7 @@ namespace mu2e{
        << _allProcesses.size() << " "
        << endl;
 
-    int tcsum(0);
+    size_t tcsum(0);
     for ( map_type::const_iterator i=_allProcesses.begin();
           i != _allProcesses.end(); ++i ){
       ProcInfo const& oi        = i->second;
